ispravljena greška u validaciji

+++ src/hr/josip/mydata/processing/PersonalId.java
@@ -43,11 +43,7 @@ public class PersonalId {
         this.dateOfIssue = dateOfIssue;
         this.personalIdentificationNumber = personalIdentificationNumber;
     }
-    //This function checks if the string is numeric
-    //there is a better written function at this class
-    //it only parse integer and calls an error if
-    //no numeric value is written to object
-    //this is unnessesary but it will stay like this
+    //This function checks if the string is numeric integer
     public boolean parseString(String string){
         boolean isString=true;
         parseString=new int[string.length()];
@@ -202,7 +198,7 @@ public class PersonalId {
 
 
             try {
-                int number=Integer.parseInt(string);
+                Long number=Long.parseLong(string);
             } catch (NumberFormatException e) {
                 errors=true;
                 System.out.println("This value is not numeric!!!");
